+++ stories/mocked/disputeManagement.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta } from '@storybook/preact';
 import DisputeManagementElement from '../../src/components/external/DisputesManagement/DisputeManagementElement';
-import { DisputeManagementMeta } from '../components/disputesManagement';
+import { DisputeManagementMeta } from '../components/disputeManagement';
 import { CUSTOM_URL_EXAMPLE } from '../utils/constants';
 import { ElementProps, ElementStory } from '../utils/types';
